requirements.txt
annotated-types==0.7.0
anyio==4.6.1
arrow==1.3.0
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.8.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.115.2
fastjsonschema==2.20.0
Flask==3.0.3
Flask-MySQLdb==2.0.0
flet-core==0.24.1
flet-runtime==0.24.1
fonttools==4.54.1
frozendict==2.4.6
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
ibge==0.0.5
idna==3.10
ipeadatapy==0.1.9
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multitasking==0.0.11
mysql-connector-python==9.1.0
mysqlclient==2.2.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.1.2
oauthlib==3.2.2
openpyxl==3.1.5
packaging==23.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
peewee==3.17.7
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.2.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
pywin32==307
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==7.4.2
referencing==0.35.1
repath==0.9.0
requests==2.32.3
retrying==1.3.4
rich==13.9.2
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
sgs==2.1.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.39.2
statsmodels==0.14.4
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
watchdog==4.0.2
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.4
yfinance==0.2.48
